Adding Vagrant
Set up to install apache2, MySQL and php.
Just open a command prompt in the directory that has the Vagrantfile and run "vagrant up"
You should be able to open the website at localhost:4567
test, via ssh or however, that you have MySQL and php installed. Although we can decide to use python, which comes preinstalled with ubuntu.
Remember that you need a few GBs of space